Move internal/x/cmd to internal/cmd
