+++ internal/cmd/testdata/compile/errors_on_import.proto
