+++ internal/cmd/testdata/compile/json_camel_case_conflict.proto
