+++ internal/cmd/testdata/compile/missing_package_semicolon.proto
