+++ internal/cmd/testdata/compile/missing_syntax.proto
