+++ internal/cmd/testdata/compile/prototool.yaml
