+++ internal/cmd/testdata/compile/syntax_proto2.proto
