+++ internal/cmd/testdata/format/bar/bar.proto
