+++ internal/cmd/testdata/format/foo/foo_proto2.proto
