+++ internal/cmd/testdata/format/prototool.yaml
