+++ internal/cmd/testdata/lint/samedir/foo2.proto
