+++ internal/cmd/testdata/lint/samedirgopkg/prototool.yaml
