+++ internal/cmd/testdata/lint/samedirjavapkg/foo2.proto
